Dont aggregate school if we're not showing data enabled features
